+++ src/ArduinoJson/Array/ArrayRef.hpp
@@ -34,6 +34,10 @@ class ArrayRefBase {
     return VariantConstRef(_data ? _data->get(index) : 0);
   }
 
+  FORCE_INLINE size_t memoryUsage() const {
+    return _data ? _data->memoryUsage() : 0;
+  }
+
   FORCE_INLINE size_t size() const {
     return _data ? _data->size() : 0;
   }