+++ src/ArduinoJson/Object/ObjectRef.hpp
@@ -42,6 +42,10 @@ class ObjectRefBase {
     return _data == 0;
   }
 
+  FORCE_INLINE size_t memoryUsage() const {
+    return _data ? _data->memoryUsage() : 0;
+  }
+
   FORCE_INLINE size_t size() const {
     return _data ? _data->size() : 0;
   }